scheduled_call_data/requirements.txt
requests
python-dotenv
boto3
pandas
pyarrow
fastparquet
pyiceberg[s3fs,sql-sqlite]